+++ flutter_angle/CHANGELOG.md
@@ -1,7 +1,11 @@
-## 0.0.1
+## 0.0.3
 
-* Initial release.
+* Fixed windows not loading properly.
 
 ## 0.0.2
 
 * Updated readme, edited description, and added web to yaml.
+
+## 0.0.1
+
+* Initial release.