+++ flutter_angle_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Fixed windows not loading properly.
+
 ## 0.0.1
 
 * Initial release.